app/requirements.txt
matplotlib
pandas
seaborn
flask
flask-login
flask_mysqldb
mysql-connector-python